fix: corrected naming, change PlaceHolder to Placeholder
