+++ src/lib/edra/headless/components/ImagePlaceholder.svelte
