+++ src/lib/edra/headless/components/VideoPlaceholder.svelte
