+++ src/lib/edra/shadcn/components/ImagePlaceholder.svelte
